requirements.txt
shapely
numpy
Pillow
tqdm
